etc
안 쓰는 내용, 미구현 사항 제거
readme 작성
