+++ README.md
@@ -1,27 +1,119 @@
+<img src="https://img.shields.io/github/contributors/wafflestudio/22-5-team8-web?color=yellow">
+<img src="https://img.shields.io/github/commit-activity/t/wafflestudio/22-5-team8-web">
+<img src="https://img.shields.io/github/issues-pr/wafflestudio/22-5-team8-web?color=deep%20green">
+<img src="https://img.shields.io/github/issues-pr-closed/wafflestudio/22-5-team8-web?color=violet">
+<br><br>
+
+<div align="center">
+<h1>🧇Wachapedia</h1>
+<h4>WaffleStudio 22-5 Team 8 토이프로젝트<h4>
+</div>
+
 # 22-5 와챠피디아 Frontend
 
-- 와챠피디아(WachaPedia)는 왓챠피디아(WatchaPedia)를 클론하는 클론 프로젝트입니다.
-- 왓챠피디아는 영화, 드라마, 웹툰 등 다양한 컨텐츠에 대한 자신의 평가 및 감상을 공유하는 서비스입니다.
-- 이 레포지토리는 Frontend 레포지토리로, Backend 내용은 Backend 레포지토리를 참고하시기 바랍니다.
+- [**와챠피디아(WachaPedia)**](https://d2vsqxcvld4zf7.cloudfront.net/)는 [**왓챠피디아(WatchaPedia)**](https://pedia.watcha.com/ko-KR)를 클론하는 클론 프로젝트입니다.
+- 왓챠피디아는 본래 영화, 드라마, 웹툰 등 다양한 컨텐츠에 대한 자신의 평가 및 감상을 공유하는 서비스로, 이 프로젝트에서는 영화에 대해서만 구현하였습니다.
+- 이 레포지토리는 Frontend 레포지토리로, Backend 내용은 [Backend 레포지토리](https://github.com/wafflestudio/22-5-team8-server)를 참고하시기 바랍니다.
 
 ## 배포
 
-- AWS S3, Cloudfront를 사용했습니다.
-
-## 기술 스택
-
-- React, Typescript, Tailwind CSS
-
-## 주요 페이지 및 기능
-
-- TODO
-
-## 신기능
-
-- TODO
+- Frontend prod 배포: https://d2vsqxcvld4zf7.cloudfront.net/
+- Frontend dev 배포: https://d1xtmfti6ypide.cloudfront.net/
+- 백엔드 배포: http://52.79.121.53
 
 ## 역할 분담
 
-- 박인혁: 회원가입 및 로그인, 유저 프로필 페이지, 캘린더 구현
-- 박종경: 랜딩 페이지, 검색, 컬렉션 관련 기능, 추천 영화 페이지 구현
-- 허태윤: 영화 페이지, 리뷰, 댓글, 취향 분석 페이지 구현
+- 박인혁: 회원가입 및 로그인, 유저 프로필 페이지 구현
+- 박종경: 랜딩 페이지, 검색, 컬렉션 관련 기능, 추천 영화 페이지, 소셜 로그인, 유저 프로필 관련 기능 구현
+- 허태윤: 영화 페이지, 리뷰, 댓글, 취향 분석 페이지, 유저 차단 관련 기능 구현
+
+## 기술 스택
+
+<div>
+<img src="https://img.shields.io/badge/React-61DAFB?style=for-the-badge&logo=react&logoColor=white">
+<img src="https://img.shields.io/badge/typescript-3178C6?style=for-the-badge&logo=typescript&logoColor=white">
+<img src="https://img.shields.io/badge/Tailwind_CSS-grey?style=for-the-badge&logo=tailwind-css&logoColor=38B2AC">
+</div>
+
+## 협업
+
+<div>
+<img src="https://img.shields.io/badge/Slack-4A154B?style=for-the-badge&logo=Slack&logoColor=white">
+<img src="https://img.shields.io/badge/Notion-000000?style=for-the-badge&logo=Notion&logoColor=white">
+</div>
+
+## 주요 페이지 및 기능
+
+<img src="https://github.com/user-attachments/assets/cb4832a7-1e32-4574-81ff-a66d7133aaff" width="32%">
+<img src="https://github.com/user-attachments/assets/8954dd77-51c0-4804-9024-26bbb9fcc715" width="32%">
+<br>
+<li> 메인 페이지: 왓챠피디아 HOT 랭킹, 박스오피스 순위, 왓챠피디아 구매 순위, 넷플릭스 순위 등의 영화 목록을 제공합니다.
+
+<img src="https://github.com/user-attachments/assets/27469c0c-ca94-4f63-8bd5-3eb1ccaf2815" width="32%">
+<img src="https://github.com/user-attachments/assets/0b4d94e7-e0aa-4b9e-ac02-72ad16a74858" width="32%">
+<br>
+<li> 로그인 및 회원가입: 카카오, 구글 소셜 로그인을 지원합니다.
+
+<img src="https://github.com/user-attachments/assets/142e32a4-1a50-4a52-b33b-49547fc85bd3" width="32%">
+<img src="https://github.com/user-attachments/assets/8bd315d5-b641-4316-8e3a-3066fdedaff2" width="32%">
+<img src="https://github.com/user-attachments/assets/491a221a-5aa8-4bc0-a3ec-f9505a240ddd" width="32%">
+<br>
+<li> 영화 페이지: 영화별로 영화에 대한 포스터, 줄거리, 출연/제작 등의 정보를 제공하고, 별점을 매기고 자신의 의견을 코멘트로 작성할 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/e3e8c716-bf23-4ee0-86a8-b9010c87d14c" width="32%">
+<img src="https://github.com/user-attachments/assets/696c4eb2-cb8e-4a1e-8609-a33a2c3d0035" width="32%">
+<img src="https://github.com/user-attachments/assets/0209b920-361c-4c15-a1b5-dbff980945ce" width="32%">
+<br>
+<li> 코멘트 페이지: 코멘트를 남기고, 코멘트 목록을 볼 수 있습니다. 팔로우하는 유저의 코멘트만 모아서 볼 수 있고, 코멘트 별로 코멘트 페이지에서 댓글을 남길 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/48c29c99-55f1-4845-8150-4dbdbfb913f0" width="32%">
+<img src="https://github.com/user-attachments/assets/3b0c0fda-ff00-487f-90c4-e03741e8e140" width="32%">
+<br>
+<li> 검색 페이지: 영화, 장르, 인물, 컬렉션, 유저에 대한 통합 검색이 가능합니다.
+
+<img src="https://github.com/user-attachments/assets/25004481-773b-4dc3-9c44-c2de26b988bb" width="32%">
+<img src="https://github.com/user-attachments/assets/bd6a14ed-6f5d-4f42-aab9-9228462f9a87" width="32%">
+<br>
+<img src="https://github.com/user-attachments/assets/2aeb7ba0-f998-4e8b-8ce3-db2dcb9765ec" width="32%">
+<img src="https://github.com/user-attachments/assets/25c30000-2fa2-48ec-a73d-f877685489ae" width="32%">
+<br>
+<li> 유저 페이지: 사용자에 대한 기본 정보를 제공합니다. 팔로워/팔로잉 목록, 평가/컬렉션 목록을 확인할 수 있습니다. 로그인 시 별점 분포, 나의 좋아요 인물/컬렉션/코멘트 목록을 확인할 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/856a827e-576c-4714-a614-82cf00efd0e6" width="32%">
+<img src="https://github.com/user-attachments/assets/cba1ba53-bafb-46cc-aa9a-3d4fd0f53356" width="32%">
+<img src="https://github.com/user-attachments/assets/94c82ab0-f161-46a2-bd98-9d86f24bc3da" width="32%">
+<br>
+<img src="https://github.com/user-attachments/assets/e6a5ec2b-8ff0-4728-9ae0-80a4d6904204" width="32%">
+<img src="https://github.com/user-attachments/assets/f198a5e3-f10a-49f2-94e0-79d2e3e56b52" width="32%">
+<img src="https://github.com/user-attachments/assets/8e9a1f73-64b5-4a9c-a84c-9db54f40d9a6" width="32%">
+<br>
+<li> 컬렉션 페이지: 영화를 선택하고 편집하여 컬렉션을 생성할 수 있습니다. 컬렉션 페이지에서는 좋아요을 누르고 댓글을 남길 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/d0c0778d-dfa0-44ea-bd67-130beda12b9d" width="32%">
+<img src="https://github.com/user-attachments/assets/12869b34-f5ee-4551-9d1b-4d414c19e3a9" width="32%">
+<img src="https://github.com/user-attachments/assets/04893b4a-b091-48a7-a220-4aae5b7a7538" width="32%">
+<br>
+<li> 영화 페이지에서 컬렉션 생성: 더보기를 눌러 이미 존재하는 컬렉션에 영화를 추가하거나, 이 영화를 포함하여 새로운 컬렉션을 생성할 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/58c93740-eced-420c-9cb6-bc206ef1dcd6" width="32%">
+<img src="https://github.com/user-attachments/assets/ee05db27-ee9c-43c7-9da9-ef155746b81e" width="32%">
+<br>
+<li> 취향 분석 페이지: 사용자가 남긴 평가를 바탕으로 별점 분포, 영화 선호 태그, 영화 선호 장르, 선호 배우, 선호 감독, 영화 감상 시간 등의 통계를 확인할 수 있습니다.
+
+<img src="https://github.com/user-attachments/assets/6ab9a2e2-4bb7-4e53-aef5-b9a9a2ef02fa" width="32%">
+<img src="https://github.com/user-attachments/assets/e7da5916-917a-4854-974c-612c76f3fbde" width="32%">
+<br>
+<li> 추천 페이지: 사용자가 남긴 평가를 분석하여 사용자에게 맞춤형 추천과 예상 별점을 제공합니다.
+
+## 신기능
+
+<img src="https://github.com/user-attachments/assets/b4bb17a6-2294-4f18-8cef-7b69005cc6c2" width="32%">
+<img src="https://github.com/user-attachments/assets/50ef6b4b-9289-466a-b1da-6ccbb5287878" width="32%">
+<br>
+<li> 장르 검색: 장르명을 검색하면 해당 장르의 영화를 검색할 수 있습니다. 여러 장르를 동시에 검색하는 것도 가능합니다.
+
+<img src="https://github.com/user-attachments/assets/62e7a12d-7020-459d-9101-fbd020c41766" width="32%">
+<img src="https://github.com/user-attachments/assets/e02eb289-4843-4601-9d9f-07d37711cfc7" width="32%">
+<img src="https://github.com/user-attachments/assets/235a4cd0-bba4-4e15-9cb0-77381f93ee2b" width="32%">
+<br>
+<li> 사용자 차단: 사용자 페이지에서 차단할 수 있고, 차단하면 영화 리뷰나 검색 결과에서 해당 사용자를 확인할 수 없습니다. 차단 관리 페이지에서 다시 차단을 해제할 수 있습니다.
