+++ package.json
@@ -16,10 +16,8 @@
   },
   "dependencies": {
     "autoprefixer": "10.4.20",
-    "moment": "^2.30.1",
     "postcss": "8.4.49",
     "react": "18.3.1",
-    "react-calendar": "^5.1.0",
     "react-device-detect": "^2.2.3",
     "react-dom": "18.3.1",
     "react-icons": "^5.4.0",
